split lateral tests
